refactor (ontology): move entity Iri handling to utils
