+++ src/api/v2/ontology/ontologies-endpoint.spec.ts
@@ -5,6 +5,7 @@ import { ReadOntology } from "../../../models/v2/ontologies/read-ontology";
 import { ResourceClassDefinition } from "../../../models/v2/ontologies/resource-class-definition";
 import { ResourcePropertyDefinition } from "../../../models/v2/ontologies/resource-property-definition";
 import { SystemPropertyDefinition } from "../../../models/v2/ontologies/system-property-definition";
+import { OntologyConversionUtils } from "../../../models/v2/OntologyConversionUtil";
 
 describe("OntologiesEndpoint", () => {
 
@@ -21,21 +22,21 @@ describe("OntologiesEndpoint", () => {
 
     describe("Method getOntologyIriFromEntityIri", () => {
         it("should get the ontology IRI from a knora-api entity", () => {
-            const ontologyIri = knoraApiConnection.v2.onto.getOntologyIriFromEntityIri("http://api.knora.org/ontology/knora-api/v2#Resource");
+            const ontologyIri = OntologyConversionUtils.getOntologyIriFromEntityIri("http://api.knora.org/ontology/knora-api/v2#Resource", config);
 
             expect(ontologyIri[0]).toEqual("http://api.knora.org/ontology/knora-api/v2");
 
         });
 
         it("should get the ontology IRI from a project entity", () => {
-            const ontologyIri = knoraApiConnection.v2.onto.getOntologyIriFromEntityIri("http://api.dasch.swiss/ontology/0807/mls/v2#Article");
+            const ontologyIri = OntologyConversionUtils.getOntologyIriFromEntityIri("http://api.dasch.swiss/ontology/0807/mls/v2#Article", config);
 
             expect(ontologyIri[0]).toEqual("http://api.dasch.swiss/ontology/0807/mls/v2");
 
         });
 
         it("should ignore an external entity", () => {
-            const ontologyIri = knoraApiConnection.v2.onto.getOntologyIriFromEntityIri("http://www.w3.org/2000/01/rdf-schema#label");
+            const ontologyIri = OntologyConversionUtils.getOntologyIriFromEntityIri("http://www.w3.org/2000/01/rdf-schema#label", config);
 
             expect(ontologyIri.length).toEqual(0);
